xive/p9: Allocate all XIVE structures on a selected chip
Currently, the NVTs and ENDs are distributed across the chips of the
system. For test and performance, it is interesting to have a better
control on the allocation of the XIVE structures. Introduce a way to
force the allocation of the NVTs, the associated ENDs and the IPIs on
a XIVE preferred chip.

The NVTs and associated ENDs of the HW threads are still allocated on
their respective chip but for all OS interrupt contexts, the XIVE
structures are allocated on the preferred chip.

This behavior can be changed with the debugfs handler.

Signed-off-by: Cédric Le Goater <clg@kaod.org>